Update Resume/Nav to Reflect Site Changes
Modified link to be used to go to resume. Previous one was broken. Decided to let it link directly to Google Drive instead to avoid errors in the future.
Modified display of the sliding navbar. Text was not aligned correctly. Tried multiple methods and elected to transform: translateX(1.25rem).... this solved the problem without causing any major issues.